requirements.txt
protobuf==6.30.2
python-dotenv==1.1.0
google-genai==1.13.0